[#52] fix fast refresh linter erros by moving affected parts to separate files
also fix base page to extend height if content overflows
