+++ frontend/app/src/components/ui/input.tsx
@@ -2,10 +2,7 @@ import * as React from "react";
 
 import { cn } from "@/lib/utils";
 
-//components from the ui folder are official shadcn components, their warnings can therefore be ignored
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface InputProps
-	extends React.InputHTMLAttributes<HTMLInputElement> {}
+export type InputProps = React.InputHTMLAttributes<HTMLInputElement>;
 
 const Input = React.forwardRef<HTMLInputElement, InputProps>(
 	({ className, type, ...props }, ref) => {
