+++ frontend/app/src/components/ui/textarea.tsx
@@ -2,10 +2,7 @@ import * as React from "react";
 
 import { cn } from "@/lib/utils";
 
-//components from the ui folder are official shadcn components, their warnings can therefore be ignored
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface TextareaProps
-	extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {}
+export type TextareaProps = React.TextareaHTMLAttributes<HTMLTextAreaElement>;
 
 const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
 	({ className, ...props }, ref) => {
