+++ frontend/app/src/services/userContext/UserContext.tsx
@@ -1,12 +1,9 @@
-import {
-	Context,
-	createContext,
-	Dispatch,
-	ReactNode,
-	useContext,
-	useReducer,
-} from "react";
+import { Context, createContext, Dispatch, ReactNode, useReducer } from "react";
 import { User } from "@/domain/user.ts";
+import {
+	playerA,
+	userReducer,
+} from "@/services/userContext/UserContextHelpers.ts";
 
 export const UserContext: Context<User | null> = createContext<User | null>(
 	null,
@@ -30,38 +27,3 @@ export function UserProvider({ children }: { children: ReactNode }) {
 		</UserContext.Provider>
 	);
 }
-// See https://react.dev/learn/scaling-up-with-reducer-and-context, they suggest this there
-// eslint-disable-next-line react-refresh/only-export-components
-export function useUser() {
-	return useContext(UserContext);
-}
-
-// See https://react.dev/learn/scaling-up-with-reducer-and-context, they suggest this there
-// eslint-disable-next-line react-refresh/only-export-components
-export function useDispatchUser() {
-	return useContext(UserDispatchContext);
-}
-
-// the user field in action is currently nullable and pretty much everywhere null, but I think we need it for the reducer,
-// and since we will need it for the future, we are just gonna keep it :3
-// Export for testing
-// eslint-disable-next-line react-refresh/only-export-components
-export function userReducer(
-	_user: User | null,
-	action: { type: string; user: User | null },
-) {
-	switch (action.type) {
-		case "setPlayerA": {
-			return playerA;
-		}
-		case "setPlayerB": {
-			return playerB;
-		}
-		default: {
-			throw Error("Unknown action: " + action.type);
-		}
-	}
-}
-
-const playerA: User = new User("PlayerA", "#FF0000");
-const playerB: User = new User("PlayerB", "#0000FF");
